requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
click==8.1.7
distro==1.9.0
exceptiongroup==1.3.1
Flask==3.1.2
Flask-WTF==1.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
injector==0.23.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
openai==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.4
WTForms==3.2.1

pytest
flask-sqlalchemy
flask-migrate
langchain-core
langchain-openai
langchain